Added RequestHistory page with Modified RoleRequest and Router also set link of RequestHistory in the nav-main component
