+++ src/pages/Patient/RequestForRole/RoleRequest.jsx
@@ -1,93 +1,8 @@
-import {
-  Table,
-  TableBody,
-  TableCaption,
-  TableCell,
-  TableFooter,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "@/components/ui/table";
 import RequestForm from "@/pages/RequestForm/RequestForm";
-import axios from "axios";
-import { CopyX, CornerUpRight } from "lucide-react";
-import Swal from "sweetalert2";
-import toast from "react-hot-toast";
-import { useRoleRequest } from "@/hooks/useRoleRequest";
-import { useEffect, useState } from "react";
-import RequestTableRow from "@/components/RequestTableRow/RequestTableRow";
+import { CornerUpRight } from "lucide-react";
 
 const RoleRequest = () => {
-  const [requestedData, refetch] = useRoleRequest();
-  const [isLoading, setIsLoading] = useState(true);
-  const [requestModal, setRequestModal] = useState({});
-  const handleView = (request) => {
-    // console.log(request);
-    setRequestModal(request);
-    document.getElementById("request_modal").showModal();
-  };
-
-  const handleCancelRequest = async (id) => {
-    // console.log(id);
-    setIsLoading(true);
-
-    Swal.fire({
-      title: "Are you sure?",
-      text: "You will cancel this request!",
-      icon: "warning",
-      showCancelButton: true,
-      confirmButtonColor: "#3085d6",
-      cancelButtonColor: "#d33",
-      confirmButtonText: "Yes, cancel!",
-    }).then(async (result) => {
-      if (result.isConfirmed) {
-        const response = await axios.patch(
-          `${import.meta.env.VITE_API_URL}/user-requests/status/${id}`
-        );
-
-        if (response.status === 200) {
-          refetch();
-          setIsLoading(false);
-          toast.success("Status has been updated");
-        }
-      }
-    });
-  };
-
-  const handleDeleteRequest = async (id) => {
-    setIsLoading(true);
-
-    Swal.fire({
-      title: "Are you sure?",
-      text: "You cannot retrieve this request!",
-      icon: "warning",
-      showCancelButton: true,
-      confirmButtonColor: "#3085d6",
-      cancelButtonColor: "#d33",
-      confirmButtonText: "Yes, delete!",
-    }).then(async (result) => {
-      if (result.isConfirmed) {
-        const response = await axios.delete(
-          `${import.meta.env.VITE_API_URL}/user-requests/${id}`
-        );
-        if (response.status === 200) {
-          refetch();
-          setIsLoading(false);
-          toast.success("Request has been deleted");
-        }
-      }
-    });
-  };
-
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setIsLoading(false);
-    }, 2000);
-
-    // Cleanup
-    return () => clearTimeout(timer);
-  }, []);
-
+  
   return (
     <div className="space-y-4 lg:w-full w-11/12 mx-auto">
       {/* Heading */}
@@ -105,181 +20,6 @@ const RoleRequest = () => {
 
       {/* Request Form */}
       <RequestForm />
-
-      {/* Requested Table Data */}
-      <div className="py-8 rounded-xl">
-        <Table
-          className={
-            "*:w-full *:rounded-xl border border-gray-200 dark:border-gray-700"
-          }
-        >
-          <TableCaption>A list of your upgrade requests.</TableCaption>
-          <TableHeader>
-            <TableRow>
-              <TableHead>Serial</TableHead>
-              <TableHead>Photo</TableHead>
-              <TableHead>Name</TableHead>
-              <TableHead>Email</TableHead>
-              <TableHead>Contact</TableHead>
-              <TableHead>Shift</TableHead>
-              <TableHead>Request Date</TableHead>
-              <TableHead>Department</TableHead>
-              <TableHead>Request Role</TableHead>
-              <TableHead>Status</TableHead>
-              <TableHead>Actions</TableHead>
-            </TableRow>
-          </TableHeader>
-          <TableBody>
-            {isLoading
-              ? Array.from({ length: 5 }).map((_, index) => (
-                  <TableRow
-                    key={index}
-                    className="hover:bg-gray-100 dark:hover:bg-gray-700"
-                  >
-                    <TableCell>
-                      <div className="skeleton h-4 w-4" />
-                    </TableCell>
-                    <TableCell>
-                      <div className="skeleton h-12 w-12 rounded-full" />
-                    </TableCell>
-                    <TableCell>
-                      <div className="skeleton h-4 w-32" />
-                    </TableCell>
-                    <TableCell>
-                      <div className="skeleton h-4 w-48" />
-                    </TableCell>
-                    <TableCell>
-                      <div className="skeleton h-4 w-24" />
-                    </TableCell>
-                    <TableCell>
-                      <div className="skeleton h-4 w-20" />
-                    </TableCell>
-                    <TableCell>
-                      <div className="skeleton h-4 w-28" />
-                    </TableCell>
-                    <TableCell>
-                      <div className="skeleton h-4 w-32" />
-                    </TableCell>
-                    <TableCell>
-                      <div className="skeleton h-4 w-24" />
-                    </TableCell>
-                    <TableCell>
-                      <div className="skeleton h-4 w-16" />
-                    </TableCell>
-                    <TableCell>
-                      <div className="skeleton h-8 w-8" />
-                    </TableCell>
-                  </TableRow>
-                ))
-              : requestedData.map((request, index) => (
-                 <RequestTableRow key={request?._id || index} request={request} index={index} handleCancelRequest={handleCancelRequest} handleDeleteRequest={handleDeleteRequest} handleView={handleView} />
-                ))}
-          </TableBody>
-          <TableFooter>
-            <TableRow>
-              <TableCell colSpan={2}>Total Requests:</TableCell>
-              <TableCell>
-                {isLoading ? (
-                  <div className="skeleton w-8 h-4"></div>
-                ) : (
-                  requestedData.length
-                )}
-              </TableCell>
-            </TableRow>
-          </TableFooter>
-        </Table>
-      </div>
-
-      <dialog id="request_modal" className="modal modal-middle">
-        {requestModal && (
-          <div className="w-full flex justify-center items-center">
-            <div className="modal-box">
-              <h2 className="md:text-3xl text-2xl font-bold text-center">Details Of Upgrade Request</h2>
-              <div className="divider md:w-11/12 mx-auto"></div>
-
-              <figure className="w-44 h-44 mx-auto mt-3">
-                <img
-                  className="w-full h-full border-4 border-muted overflow-hidden rounded-full object-cover"
-                  src={requestModal?.userPhoto}
-                  alt={requestModal?.userName}
-                />
-              </figure>
-
-              <div className="divider"></div>
-
-              <div className="w-full space-y-3">
-                <h4 className="text-lg text-gray-900 font-bold">
-                  Name:{" "}
-                  <span className="text-gray-700 font-semibold">
-                    {requestModal?.userName}
-                  </span>
-                </h4>
-
-                <h4 className="text-lg text-gray-900 font-bold">
-                  Email:{" "}
-                  <span className="text-gray-700 font-semibold">
-                    {requestModal?.userEmail}
-                  </span>
-                </h4>
-                
-                <h4 className="text-lg text-gray-900 font-bold">
-                  Request Role:{" "}
-                  <span className="text-gray-700 font-semibold">
-                    {requestModal?.requestedRole}
-                  </span>
-                </h4>
-
-                <h4 className="text-lg text-gray-900 font-bold">
-                  Emergency Contact:{" "}
-                  <span className="text-gray-700 font-semibold">
-                    {requestModal?.emergencyContact}
-                  </span>
-                </h4>
-
-                <h4 className="text-lg text-gray-900 font-bold">
-                  Shift:{" "}
-                  <span className="text-gray-700 font-semibold">
-                    {requestModal?.shift}
-                  </span>
-                </h4>
-
-                <h4 className="text-lg text-gray-900 font-bold">
-                  Available Moment:{" "}
-                  <span className="text-gray-700 font-semibold">
-                    {new Date(requestModal?.availableDate).toLocaleString("en-UK")}
-                  </span>
-                </h4>
-
-                <h4 className="text-lg text-gray-900 font-bold">
-                  Address:{" "}
-                  <span className="text-gray-700 font-semibold">
-                    {requestModal?.address}
-                  </span>
-                </h4>
-
-                <h4 className="text-lg text-gray-900 font-bold">
-                  Cover Letter:{" "}
-                  <span className="text-gray-700 font-semibold">
-                    {requestModal?.coverLetter}
-                  </span>
-                </h4>
-
-                <div className="py-3">
-                  <button
-                    onClick={() =>
-                      document.getElementById("request_modal").close()
-                    }
-                    className="md:px-14 px-10 btn bg-rose-500 text-base text-white font-bold rounded-md flex gap-2 items-center"
-                  >
-                    <CopyX className="w-4 h-4" />
-                    <span>Close</span>
-                  </button>
-                </div>
-              </div>
-            </div>
-          </div>
-        )}
-      </dialog>
     </div>
   );
 };
